% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineAndDiscard.R
\name{Combine_and_Discard}
\alias{Combine_and_Discard}
\title{Combine and Discard}
\usage{
Combine_and_Discard(merged_laels, reference = 1, fraction = 0.05, verbose = T)
}
\arguments{
\item{reference}{the index of reference to be based on}

\item{fraction}{default is 0.05}

\item{merged_labels}{merged_labels got from \code{get_merged_labels()}}
}
\value{
Return the index of cells to be discarded.
}
\description{
Combine cell labels predicted by each reference and discard a fraction of cells with uncertain labels
}
