% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPLabelCorrector.R
\name{correct_labels}
\alias{correct_labels}
\title{Correct labels}
\usage{
correct_labels(
  merged_labels,
  query_nmf_embedding,
  cutoff = 0.6,
  max_edit_distance = NULL,
  threads = 10,
  verbose = TRUE
)
}
\arguments{
\item{merged_labels}{merged_labels got from \code{get_merged_labels()}}

\item{query_nmf_embedding}{Query meta-program matrix}

\item{cutoff}{The cutoff to drfinr 'similar' cells, smaller value means more strict correction}

\item{max_edit_distance}{Max edit distance, by default is the number of references-1}

\item{threads}{The number of threads}

\item{verbose}{Whether to display messages}
}
\value{
Returns a list containing the information of the corrected labels
}
\description{
Correct labels
}
