% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_all_markers}
\alias{get_all_markers}
\title{Get all markers}
\usage{
get_all_markers(
  pb.ref,
  atlas = "H",
  cell.subtype,
  HEG.prop = 0.2,
  marker.num = 200
)
}
\arguments{
\item{pb.ref}{a pseudo bulk matrix whose rows are features and columns are cell types}

\item{atlas}{which atlas to use, default is \code{'H'}
\itemize{
\item{H: }{use the Human Cell Atlas}
\item{M: }{use the Mouse Cell Atlas}
}}

\item{cell.subtype}{the list of cell subtypes get from \code{get_cell_subtype()}}

\item{HEG.prop}{the proportion of highly expressed genes to first find, default is 0.2}

\item{marker.num}{the number of markers to find of each cell type}
}
\value{
Returns a list containing global markers and neighbor specific markers.
}
\description{
Get global markers and neighbor specific markers from reference
}
