% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{get_benchmark}
\alias{get_benchmark}
\title{Get benchmark indicators}
\usage{
get_benchmark(true.labels, pred.labels)
}
\arguments{
\item{true.labels}{a vector of the true labels}

\item{pred.labels}{a vector of the predicted labels}
}
\value{
Returns a vector of the accuracy, average recall (balanced accuracy), average precision and mean F1 score.
}
\description{
Get benchmark indicators
}
