% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_candidate_seed_barcodes}
\alias{get_candidate_seed_barcodes}
\title{get_candidate_seed_barcodes}
\usage{
get_candidate_seed_barcodes(cell_meta, threads = 10)
}
\arguments{
\item{cell_meta}{a cell metadata}

\item{threads}{number of threads}
}
\value{
the candidate seed barcodes
}
\description{
get_candidate_seed_barcodes
}
