% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_cell_subtype}
\alias{get_cell_subtype}
\title{Get cell subtypes from a pseudo bulk matrix}
\usage{
get_cell_subtype(pb.ref, min.cor = 0.6, verbose = T)
}
\arguments{
\item{pb.ref}{a pseudo bulk matrix whose rows are features and columns are cell types}

\item{min.cor}{the minimum cutoff of cell type correlation}

\item{verbose}{whether to display messages and plot, default is TRUE}
}
\value{
Returns a list of cell subtypes.
}
\description{
Get cell subtypes from a pseudo bulk matrix
}
