% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.3_GetKendallPred.R
\name{get_cor_mtx}
\alias{get_cor_mtx}
\title{Get feature selected pseudo bulk reference}
\usage{
get_cor_mtx(
  sc_count_mtx,
  labels,
  query_mtx,
  global_markers,
  query_nmf_embedding,
  threads = 10,
  verbose = T
)
}
\arguments{
\item{query_mtx}{query_mtx}

\item{global_markers}{global_markers}

\item{threads}{the number of threads, default is 10}

\item{verbose}{whether to display messages, default is TRUE}

\item{ref_metacell_Seurat}{ref_metacell_Seurat}
}
\value{
Returns a correlation matrix whose rows are query cells and columns are
reference cell types.
}
\description{
Get first round annotation correlation matrix
}
\details{
Get first round annotation correlation matrix of each query cell to each reference cell type using
Kendall correlation coefficient
}
\examples{
cor_mtx <- get_cor_mtx(feature_selected_pb_ref, feature_selected_query_mtx)
}
