% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkFunctions.R
\name{get_each_precision}
\alias{get_each_precision}
\title{Get each cell type's precision}
\usage{
get_each_precision(cell_meta, ident = "kendall_pred")
}
\arguments{
\item{cell_meta}{the cell metadata}

\item{ident}{the column name of the predicted labels}
}
\value{
Return a data.frame containing each cell type's precision
}
\description{
Get each cell type's precision
}
