% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkFunctions.R
\name{get_each_recall}
\alias{get_each_recall}
\title{Get each cell type's recall}
\usage{
get_each_recall(cell_meta, ident = "kendall_pred")
}
\arguments{
\item{cell_meta}{the cell metadata}

\item{ident}{the column name of the predicted labels}
}
\value{
Return a data frame containing each cell type's recall
}
\description{
Get each cell type's recall
}
