% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_feature_selected_pb_ref}
\alias{get_feature_selected_pb_ref}
\title{Get feature selected pseudo bulk reference}
\usage{
get_feature_selected_pb_ref(pb.ref, HEG.prop = 0.2, each.ct.top = 200)
}
\arguments{
\item{pb.ref}{a pseudo bulk matrix whose rows are features and columns are cell types}

\item{HEG.prop}{the proportion of highly expressed genes to first find, default is 0.2,
if the matrix is real bulk matrix it is recommended to set as 0.8}

\item{each.ct.top}{the number of top marker genes to find for a cell type}
}
\value{
Returns a feature selected pseudo bulk reference matrix.
}
\description{
Get feature selected pseudo bulk reference matrix to calculate cell-cell type correlation
}
\examples{
feature_selected_pb_ref <- get_feature_selected_pb_ref(pb_ref)
}
