% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_feature_selected_query_mtx}
\alias{get_feature_selected_query_mtx}
\title{Get feature selected query matrix}
\usage{
get_feature_selected_query_mtx(query.mtx, query.nmf.embedding)
}
\arguments{
\item{query.mtx}{query gene activity matrix whose rows are genes and columns are cells}

\item{query.nmf.embedding}{a dimension reduced matrix whose rows are query cells and
columns are factors (meta-programs)}
}
\value{
Returns a feature selected query matrix.
}
\description{
Get feature selected query matrix to calculate cell-cell type correlation
}
\examples{
get_feature_selected_query_mtx <- get_feature_selected_query_mtx(query_mtx, query_nmf_embedding)
}
