% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_gene_gr}
\alias{get_gene_gr}
\title{Get GRanges objects of all coding genes for a specific genome}
\usage{
get_gene_gr(genome)
}
\arguments{
\item{genome}{Name of the genome, could be 'hg38', 'hg19', 'mm10' or 'mm9', default is 'hg38'.}
}
\value{
Return a GRanges objects of all coding genes for the genome
}
\description{
Get GRanges objects of all coding genes for a specific genome
}
\examples{
hg38_gene_gr <- get_gene_gr('hg38')
}
