% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_global_markers_bulk}
\alias{get_global_markers_bulk}
\title{Get global markers for bulk RNA-seq gene counts matrix}
\usage{
get_global_markers_bulk(
  sc_counts_mtx,
  labels,
  max_marker = 200,
  threads = 10,
  return_raw = F
)
}
\arguments{
\item{sc_counts_mtx}{bulk RNA-seq counts matrix}

\item{labels}{bulk RNA-seq sample labels}

\item{max_marker}{maximum number of markers for each cell type}

\item{threads}{the number of threads, default is 10}

\item{return_raw}{whether return raw data, default is FALSE}
}
\value{
Returns a list of global markers for each cell type
}
\description{
Get global markers for bulk RNA-seq gene counts matrix
}
