% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_global_markers_pb}
\alias{get_global_markers_pb}
\title{get_global_markers_pb}
\usage{
get_global_markers_pb(metacell_Seurat, threads = 10)
}
\arguments{
\item{metacell_Seurat}{todo}

\item{threads}{todo}
}
\value{
todo
}
\description{
get_global_markers_pb
}
\examples{
todo
}
