% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_MAPLabelCorrector.R
\name{get_merged_labels}
\alias{get_merged_labels}
\title{get merged labels}
\usage{
get_merged_labels(pred_df)
}
\arguments{
\item{pred_df}{}
}
\value{
Return the merged label
}
\description{
get merged labels
}
