% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_metacell_exp}
\alias{get_metacell_exp}
\title{get_metacell_exp}
\usage{
get_metacell_exp(
  count_mtx,
  labels,
  meta_cell_num = 10,
  mode = "sum",
  return_Seurat = T
)
}
\arguments{
\item{count_mtx}{todo}

\item{labels}{todo}

\item{meta_cell_num}{todo}

\item{mode}{todo}

\item{return_Seurat}{todo}
}
\value{
todo
}
\description{
get_metacell_exp
}
\examples{
todo
}
