% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_neighbor_celltypes}
\alias{get_neighbor_celltypes}
\title{Get cell neighbors from a pseudo bulk matrix}
\usage{
get_neighbor_celltypes(
  metacell_Seurat,
  global_markers,
  min.cor = 0.6,
  verbose = T
)
}
\arguments{
\item{min.cor}{the minimum cutoff of cell type correlation}

\item{verbose}{whether to display messages and plot, default is TRUE}

\item{pb.ref}{a pseudo bulk matrix whose rows are features and columns are cell types}
}
\value{
Returns a list of cell subtypes.
}
\description{
Get cell neighbors from a pseudo bulk matrix
}
