% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_neighbor_markers_bulk}
\alias{get_neighbor_markers_bulk}
\title{Get neighbor markers for bulk RNA-seq gene counts matrix}
\usage{
get_neighbor_markers_bulk(
  sc_counts_mtx,
  labels,
  neighbor_celltypes,
  global_markers,
  max_marker = 200,
  threads = 10
)
}
\arguments{
\item{sc_counts_mtx}{Bulk RNA-seq counts matrix}

\item{labels}{Bulk RNA-seq sample labels}

\item{neighbor_celltypes}{Neighbor celltypes got from \code{get_neighbor_celltypes()}}

\item{global_markers}{Global markers got from \code{get_global_markers_bulk()}}

\item{max_marker}{Maximum number of markers for each cell type}

\item{threads}{The number of threads, default is 10}
}
\value{
Returns a list of neighbor markers for each cell type
}
\description{
Get neighbor markers for bulk RNA-seq gene counts matrix
}
