% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_neighbor_markers_pb}
\alias{get_neighbor_markers_pb}
\title{get_neighbor_markers_pb}
\usage{
get_neighbor_markers_pb(
  metacell_Seurat,
  neighbor_celltypes,
  global_markers,
  threads = 10
)
}
\arguments{
\item{metacell_Seurat}{todo}

\item{neighbor_celltypes}{todo}

\item{global_markers}{todo}

\item{threads}{todo}
}
\value{
todo
}
\description{
get_neighbor_markers_pb
}
\examples{
todo
}
