% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_neighbor_markers_sc}
\alias{get_neighbor_markers_sc}
\title{Get neighbor markers for scRNA-seq gene counts matrix}
\usage{
get_neighbor_markers_sc(
  sc_counts_mtx,
  labels,
  neighbor_celltypes,
  global_markers,
  max_marker = 200,
  threads = 10
)
}
\arguments{
\item{sc_counts_mtx}{scRNA-seq counts matrix}

\item{labels}{scRNA-seq cell labels}

\item{neighbor_celltypes}{neighbor celltypes got from \code{get_neighbor_celltypes()}}

\item{global_markers}{global markers got from \code{get_global_markers_sc()}}

\item{max_marker}{maximum number of markers for each cell type}

\item{threads}{the number of threads, default is 10}
}
\value{
Returns a list of neighbor markers for each cell type
}
\description{
Get neighbor markers for scRNA-seq gene counts matrix
}
