% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{get_pb_markers}
\alias{get_pb_markers}
\title{Get markers from a pseudo bulk matrix}
\usage{
get_pb_markers(
  pb.mtx,
  ct.to.find.mk = NULL,
  HEG.prop = 0.2,
  each.ct.top = 200,
  min.cor = NULL,
  return.cor.mtx = FALSE
)
}
\arguments{
\item{pb.mtx}{a pseudo bulk matrix whose rows are features and columns are cell types}

\item{ct.to.find.mk}{a vector of cell types to find markers, default is NULL (all cell types)}

\item{HEG.prop}{the proportion of highly expressed genes to first find, default is 0.2,
if the matrix is a real bulk matrix, this parameter is recommended to set as 0.8}

\item{each.ct.top}{the number of top marker genes to find for each cell type}

\item{return.type}{the type to return, default is \code{'df'}
\itemize{
\item{df: }{a dataframe whose rows are marker genes and columns are cell types}
\item{vector: }{a vector contains marker genes of all cell types}
}}
}
\value{
If \code{return.type = 'df'}, returns a data frame whose rows are marker genes and columns are cell types.
If \code{return.type = 'vector'}, returns a vector containing marker genes of all cell types.
}
\description{
Get markers from a pseudo bulk matrix
}
