% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_seed_barcode}
\alias{get_seed_barcode}
\title{get_seed_barcode}
\usage{
get_seed_barcode(cell_meta, threads = 10)
}
\arguments{
\item{cell_meta}{a cell metadata}

\item{threads}{number of threads}
}
\value{
the seed barcodes
}
\description{
get_seed_barcode
}
