% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_seed_candidates}
\alias{get_seed_candidates}
\title{get_seed_barcode}
\usage{
get_seed_candidates(cell_meta)
}
\arguments{
\item{cell_meta}{}
}
\value{
the seed barcodes
}
\description{
get_seed_barcode
}
