% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsPlotFunctions.R
\name{plot_celltype_proportions}
\alias{plot_celltype_proportions}
\title{Plot cell type proportions for the reference, candidate seed cells, seed cells and the final predictions}
\usage{
plot_celltype_proportions(cell_meta, ref_labels)
}
\arguments{
\item{cell_meta}{The cell metadata}

\item{ref_labels}{The cell type labels of the reference cells}
}
\value{
Return a ggplot object
}
\description{
Plot cell type proportions for the reference, candidate seed cells, seed cells and the final predictions
}
