% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsPlotFunctions.R
\name{plot_highlight_cells}
\alias{plot_highlight_cells}
\title{Plot highlight cells}
\usage{
plot_highlight_cells(
  Seurat.object,
  celltype,
  ident,
  label = TRUE,
  pt.size = 1,
  reduction = SeuratObject::DefaultDimReduc(Seurat.object),
  colors = NULL
)
}
\arguments{
\item{Seurat.object}{A Seurat object}

\item{celltype}{Which cell type(s) to highlight}

\item{ident}{The identity class to plot}

\item{label}{Whether to display the cell type label on the plot}

\item{pt.size}{Point size}

\item{reduction}{Which dimensionality reduction to use}

\item{colors}{Colors of different cell types}
}
\value{
Returns a ggplot object
}
\description{
Plot highlight cells
}
