% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsPlotFunctions.R
\name{plot_pred_nmf}
\alias{plot_pred_nmf}
\title{Plot the NMF meta-program heatmap of the predictions}
\usage{
plot_pred_nmf(
  query_nmf_embedding,
  cell_meta,
  neighbor_celltypes,
  category = "final",
  sample_cells = F,
  celltypes_to_plot = NULL
)
}
\arguments{
\item{query_nmf_embedding}{The query nmf embedding}

\item{cell_meta}{The cell metadata}

\item{neighbor_celltypes}{\code{neighbor_celltypes} generate from the function \code{get_neighbor_celltypes()}}

\item{category}{Which predictions to show, either \code{'seed_candidate'}, \code{'seed'}, or \code{final}, default is \code{'final'}}

\item{sample_cells}{Whether to sample cells, default is FALSE, if is true, randomly sample 30 cells for each cell type}

\item{celltypes_to_plot}{Cell type to display, default is all cell types}
}
\value{
Returns a Heatmap-class object
}
\description{
Plot the NMF meta-program heatmap of the predictions
}
