% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsPlotFunctions.R
\name{plot_pred_umap}
\alias{plot_pred_umap}
\title{Plot the UMAP of the predictions}
\usage{
plot_pred_umap(
  Seurat.object,
  cell_meta,
  celltype = NULL,
  category = "final",
  label = T,
  pt.size = 0.3,
  reduction = NULL,
  colors = NULL
)
}
\arguments{
\item{Seurat.object}{A Seurat Object}

\item{cell_meta}{The cell metadata}

\item{celltype}{Which cell types to show}

\item{category}{Which predictions to show, either \code{seed_candidate}, \code{seed}, or \code{final}, default is \code{final}}

\item{label}{Whether to display the cell type label on the plot, default is TRUE}

\item{pt.size}{Point size, default is 1}

\item{reduction}{Which dimensionality reduction to use}

\item{colors}{Colors of different cell types}
}
\value{
Returns a ggplot object
}
\description{
Plot the UMAP of the predictions
}
