% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsPlotFunctions.R
\name{plot_ref_global_markers_heatmap}
\alias{plot_ref_global_markers_heatmap}
\title{Plot global markers heatmap}
\usage{
plot_ref_global_markers_heatmap(
  ref_mtx,
  ref_labels,
  global_markers,
  neighbor_celltypes = NULL,
  celltypes_to_plot = NULL,
  top_marker_genes_num = 20,
  sample_cells_num = 20
)
}
\arguments{
\item{ref_mtx}{The reference gene expression matrix}

\item{ref_labels}{The cell type labels of the reference cells}

\item{global_markers}{\code{global_markers} generate from the function \code{get_global_markers_()}}

\item{neighbor_celltypes}{\code{neighbor_celltypes} generate from the function \code{get_neighbor_celltypes()}}

\item{celltypes_to_plot}{Cell type to display, default is all cell types}

\item{top_marker_genes_num}{Number of top marker genes to display for each cell type]}

\item{sample_cells_num}{Number of the sampling cells to display for each cell type}
}
\description{
Plot global markers heatmap
}
