% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsPlotFunctions.R
\name{plot_ref_neighbor_markers_heatmap}
\alias{plot_ref_neighbor_markers_heatmap}
\title{Plot neighbor markers heatmap}
\usage{
plot_ref_neighbor_markers_heatmap(
  ref_mtx,
  ref_labels,
  neighbor_markers,
  neighbor_celltypes = NULL,
  celltype_to_plot,
  top_marker_genes_num = 20,
  sample_cells_num = 20
)
}
\arguments{
\item{ref_mtx}{The reference gene expression matrix}

\item{ref_labels}{The cell type labels of the reference cells}

\item{neighbor_markers}{\code{neighbor_markers} generate from the function \code{get_neighbor_markers_()}}

\item{neighbor_celltypes}{\code{neighbor_celltypes} generate from the function \code{get_neighbor_celltypes()}}

\item{top_marker_genes_num}{Number of top marker genes to display for each cell type, default is 20}

\item{sample_cells_num}{Number of the sampling cells to display for each cell type, default is 20}

\item{celltypes_to_plot}{Which cell type and its neighbor cell types to display}
}
\description{
Plot neighbor markers heatmap
}
