% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsPlotFunctions.R
\name{plot_seed_global_markers_heatmap}
\alias{plot_seed_global_markers_heatmap}
\title{Plot global markers heatmap of seed cells gene activity}
\usage{
plot_seed_global_markers_heatmap(
  query_mtx,
  cell_meta,
  global_markers,
  neighbor_celltypes,
  celltypes_to_plot = NULL,
  top_marker_genes_num = 20,
  sample_cells_num = 20
)
}
\arguments{
\item{query_mtx}{The query gene activity matrix}

\item{cell_meta}{The cell metadata}

\item{global_markers}{\code{global_markers} generate from the function \code{get_global_markers_()}}

\item{neighbor_celltypes}{\code{neighbor_celltypes} generate from the function \code{get_neighbor_celltypes()}}

\item{celltypes_to_plot}{Cell type to display, default is all cell types}

\item{top_marker_genes_num}{Number of top marker genes to display for each cell type, default is 20}

\item{sample_cells_num}{Number of the sampling cells to display for each cell type, default is 20}
}
\description{
Plot global markers heatmap of seed cells gene activity
}
