% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.1_PreProcessing.R
\name{pre_processing}
\alias{pre_processing}
\title{Pre-process input matrices}
\usage{
pre_processing(ref_mtx, query_mtx, verbose = T)
}
\arguments{
\item{ref_mtx}{a reference gene expression matrix whose rows are genes and columns are cells/samples}

\item{query_mtx}{query gene activity matrix whose rows are genes and columns are cells}

\item{verbose}{whether to display messages, default is TRUE}
}
\value{
Returns a list containing two matrices after pre-processing.
}
\description{
Pre-process two matrices to remove 0 expressed genes and keep common genes
}
\examples{
pre_processing_mtxs <- pre_processing(ref_mtx = REF_MTX, query_mtx = GA_MTX)
ref_mtx <- pre_processing_mtxs$ref_mtx
query_mtx <- pre_processing_mtxs$query_mtx

}
