% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{RunAtacAnnoR}
\alias{RunAtacAnnoR}
\title{Run AtacAnnoR in a line}
\usage{
RunAtacAnnoR(
  ref_mtx,
  ref_celltype,
  ref_type = "sc",
  query_gene_activity,
  query_peak_counts = NULL,
  query_cell_embedding = NULL,
  min_cor = 0.6,
  num_global_markers = 200,
  num_neighbor_markers = 200,
  threads = 10,
  verbose = TRUE,
  simple_output = TRUE
)
}
\arguments{
\item{ref_mtx}{a reference gene expression counts matrix whose rows are genes and columns are cells/samples}

\item{ref_celltype}{a vector of cell types whose orders are same as the columns of \code{ref_mtx}}

\item{ref_type}{which type of the reference matrix is
\itemize{
\item{'sc': }{the reference matrix is a single cell RNA-seq matrix}
\item{'bulk': }{the reference matrix is a bulk RNA-seq matrix}
}}

\item{query_gene_activity}{a query gene activity matrix whose rows are genes and columns are cells}

\item{query_peak_counts}{a query scATAC-seq peak counts matrix whose rows are peaks and columns are cells}

\item{query_cell_embedding}{a query scATAC-seq cell embedding whose rows are cells and columns are factors,
default is NULL, which means getting NMF embedding automatically}

\item{min_cor}{the minimum of correlation to define similar cell types, default is 0.6}

\item{num_global_markers}{the number of global markers}

\item{num_neighbor_markers}{the number of neighbor markers}

\item{threads}{the number of threads, default is 10}

\item{verbose}{whether to display messages and plots, default is TRUE}

\item{simple_output}{whether to return a simple output, default is TRUE}
}
\value{
If \code{simple_output} is set to TRUE, returns a vector of predicted labels.
If \code{simple_output} is set to FALSE, returns a list object of all outputs.
}
\description{
AtacAnnoR main function
}
