% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{RunAtacAnnoR_Cicero}
\alias{RunAtacAnnoR_Cicero}
\title{Run AtacAnnoR on a cell_data_set(cds) object in Cicero pipeline.}
\usage{
RunAtacAnnoR_Cicero(
  query_cds,
  ref_mtx,
  ref_celltype,
  ref_type = "sc",
  min_cor = 0.6,
  num_global_markers = 200,
  num_neighbor_markers = 200,
  query_gene_activity,
  threads = 10,
  verbose = TRUE
)
}
\arguments{
\item{query_cds}{The query cds object}

\item{ref_mtx}{A reference gene expression matrix whose rows are genes and columns are cells/samples}

\item{ref_celltype}{A vector of cell types whose orders are same as the columns of \code{ref_mtx}}

\item{ref_type}{Which type of the reference matrix is
\itemize{
\item{'sc': }{The reference matrix is a single cell RNS-seq matrix}
\item{'bulk': }{The reference matrix is a bulk RNA-seq matrix}
}}

\item{min_cor}{The minimum of correlation to define similar cell types}

\item{num_global_markers}{The number of global markers}

\item{num_neighbor_markers}{The number of neighbor markers}

\item{query_gene_activity}{A query gene activity matrix whose rows are genes and columns are cells}

\item{threads}{The number of threads}

\item{verbose}{Whether to display messages and plots}
}
\value{
Returns a new query cds object with cell types predicted by AtacAnnoR restored in
\code{query_cds$final_pred} and NMF embedding restored in \code{query_cds@int_colData$reducedDims$NMF}.
}
\description{
Run AtacAnnoR on a cell_data_set(cds) object in Cicero pipeline. Note that before using this
function, gene activity should be calculate by Cicero (\code{https://cole-trapnell-lab.github.io/cicero-release/docs_m3/}).
}
