% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{RunAtacAnnoR_Signac}
\alias{RunAtacAnnoR_Signac}
\title{Run AtacAnnoR on a Seurat object in Signac pipeline}
\usage{
RunAtacAnnoR_Signac(
  query_SeuratObj,
  query_ga_assay = "ACTIVITY",
  query_peak_assay = "ATAC",
  ref_SeuratObj,
  ref_assay = "RNA",
  ref_ident = "celltype",
  ref_type = "sc",
  min_cor = 0.6,
  num_global_markers = 200,
  num_neighbor_markers = 200,
  threads = 10,
  verbose = TRUE
)
}
\arguments{
\item{query_SeuratObj}{The query Seurat object}

\item{query_ga_assay}{The assay containing gene activity of the query Seurat object}

\item{query_peak_assay}{The assay containing peak counts of the query Seurat object}

\item{ref_SeuratObj}{The reference Seurat object}

\item{ref_assay}{The assay of the reference Seurat object to use}

\item{ref_ident}{The identity of the reference Seurat object which contains the reference cell type}

\item{ref_type}{Which type of the reference matrix is
\itemize{
\item{'sc': }{The reference matrix is a single cell RNS-seq matrix}
\item{'bulk': }{The reference matrix is a bulk RNA-seq matrix}
}}

\item{min_cor}{The minimum of correlation to define similar cell types, default is 0.6}

\item{num_global_markers}{The number of global markers}

\item{num_neighbor_markers}{The number of neighbor markers}

\item{threads}{The number of threads, default is 10}

\item{verbose}{Whether to display messages and plots, default is TRUE}
}
\value{
Returns a new query Seurat object with cell types predicted by AtacAnnoR restored in
\code{query$final_pred} and NMF embedding restored in \code{query[['nmf']]}.
}
\description{
Run AtacAnnoR on a Seurat object in Signac pipeline.
}
