% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{RunAtacAnnoR_SnapATAC}
\alias{RunAtacAnnoR_SnapATAC}
\title{Run AtacAnnoR on a snap object in SnapATAC pipeline}
\usage{
RunAtacAnnoR_SnapATAC(
  query_snapObj,
  ref_mtx,
  ref_celltype,
  ref_type = "sc",
  min_cor = 0.6,
  num_global_markers = 200,
  num_neighbor_markers = 200,
  threads = 10,
  verbose = TRUE
)
}
\arguments{
\item{query_snapObj}{The query snap object}

\item{ref_mtx}{A reference gene expression matrix whose rows are genes and columns are cells/samples}

\item{ref_celltype}{A vector of cell types whose orders are same as the columns of \code{ref_mtx}}

\item{ref_type}{Which type of the reference matrix is
\itemize{
\item{'sc': }{The reference matrix is a single cell RNS-seq matrix}
\item{'bulk': }{The reference matrix is a bulk RNA-seq matrix}
}}

\item{min_cor}{The minimum of correlation to define similar cell types}

\item{num_global_markers}{The number of global markers}

\item{num_neighbor_markers}{The number of neighbor markers}

\item{threads}{The number of threads}

\item{verbose}{Whether to display messages and plots}
}
\value{
Returns a new query snap object with cell types predicted by AtacAnnoR restored in
\code{query_snapObj@metaData$final_pred}.
}
\description{
Run AtacAnnoR on a snap object in SnapATAC pipeline. Note that before using this
function, the \code{gmat} and \code{pmat} should be stored in
\code{query_snapObj}.
}
