% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2_SeedCleaning.R
\name{seed_cleaning}
\alias{seed_cleaning}
\title{Seed cell cleaning}
\usage{
seed_cleaning(cell_meta, query_nmf_embedding, dist.metric = "cosine")
}
\arguments{
\item{cell_meta}{a cell metadata}

\item{query_nmf_embedding}{query meta-program matrix}

\item{dist.metric}{the metric to calculate distance, can be \code{euclidean}, \code{cosine}, \code{manhattan}, and \code{hamming}, default is \code{cosine}}
}
\value{
Returns a new cell meta data with a column \code{is_seed}
}
\description{
Clean seed cell candidates
}
