% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{test_markers}
\alias{test_markers}
\title{Test markers of first round annotation}
\usage{
test_markers(
  query.mtx,
  cell.meta,
  which.label = "kendall_pred",
  all.markers,
  threads = 10,
  verbose = T
)
}
\arguments{
\item{query.mtx}{query gene activity matrix whose rows are genes and columns are cells}

\item{cell.meta}{a cell metadata}

\item{all.markers}{all markers get from \code{get_all_markers()}}

\item{threads}{the number of threads, default is 10}

\item{verbose}{whether to display messages, default is TRUE}
}
\value{
Returns a new cell metadata with columns 'GMSS','NMSS','NMFC'.
}
\description{
Test markers of first round annotation
}
