% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round2Anno.R
\name{weighted_knn}
\alias{weighted_knn}
\title{weighted_knn}
\usage{
weighted_knn(neighbors, train.labels, type = NULL)
}
\arguments{
\item{neighbors}{todo}

\item{train.labels}{todo}

\item{type}{NULL or 'prob'}
}
\value{
todo
}
\description{
weighted_knn
}
