% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round2Anno.R
\name{WKNN_predict}
\alias{WKNN_predict}
\title{Predict the label for each query cell using WKNN algorithm}
\usage{
WKNN_predict(cell_meta, query_nmf_embedding, k = 10, dist.metric = "cosine")
}
\arguments{
\item{cell_meta}{A cell meta data}

\item{query_nmf_embedding}{Query meta-program matrix}

\item{k}{The number of nearest neighbor cells to find for each query cells}

\item{dist.metric}{The metric to calculate distance, can be \code{euclidean}, \code{cosine}, \code{manhattan}, and \code{hamming}}
}
\value{
Returns a new cell meta data with a column \code{final_pred}
}
\description{
Predict the label for each query cell using WKNN algorithm
}
