% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_MAPLabelCorrector.R
\name{correct_labels}
\alias{correct_labels}
\title{correct labels}
\usage{
correct_labels(
  merged_labels,
  query_nmf_embedding,
  cutoff = 0.6,
  max_edit_distance = NULL,
  threads = 10,
  verbose = T
)
}
\arguments{
\item{merged_labels}{merged_labels got from \code{get_merged_labels()}}

\item{query_nmf_embedding}{the nmf_embedding of query data}

\item{cutoff}{0.6 by default, smaller value means more strict correction}

\item{max_edit_distance}{number of references-1 by default}

\item{threads}{number of threads}

\item{verbose}{verbose}
}
\value{
Return a list containing the information of the corrected labels
}
\description{
correct labels
}
