% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_MAPLabelCorrector.R
\name{correct_labels}
\alias{correct_labels}
\title{Correct labels}
\usage{
correct_labels(
  merged_labels,
  query_nmf_embedding,
  cutoff = 0.6,
  max_edit_distance = NULL,
  threads = 10,
  verbose = T
)
}
\arguments{
\item{merged_labels}{merged_labels got from \code{get_merged_labels()}}

\item{query_nmf_embedding}{query meta-program matrix}

\item{cutoff}{the cutoff to drfinr 'similar' cells, smaller value means more strict correction, default is 0.6}

\item{max_edit_distance}{max edit distance, by default is the number of references-1}

\item{threads}{threads the number of threads, default is 10}

\item{verbose}{whether to display messages, default is TRUE}
}
\value{
Returns a list containing the information of the corrected labels
}
\description{
Correct labels
}
