% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.3_GetKendallPred.R
\name{get_cell_meta_with_true}
\alias{get_cell_meta_with_true}
\title{Add the true labels to cell metadata}
\usage{
get_cell_meta_with_true(cell_meta, true_labels, cor_mtx = NULL)
}
\arguments{
\item{cell_meta}{a cell metadata}

\item{true_labels}{a vector of true cell labels}

\item{cor_mtx}{the correlation matrix get from \code{get_cor_mtx()}}
}
\value{
Returns a new cell metadata with true labels.
}
\description{
Add the true labels to cell metadata
}
