% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_cor_mtx}
\alias{get_cor_mtx}
\title{Get first round annotation correlation matrix}
\usage{
get_cor_mtx(
  feature.selected.pb.ref,
  feature.selected.query.mtx,
  threads = 10,
  verbose = T
)
}
\arguments{
\item{feature.selected.pb.ref}{the feature selected pseudo bulk reference}

\item{feature.selected.query.mtx}{the feature selected query matrix}

\item{threads}{the number of threads, default is 10}

\item{verbose}{whether to display messages, default is TRUE}
}
\value{
Returns a correlation matrix whose rows are query cells and columns are
reference cell types.
}
\description{
Get first round annotation correlation matrix of each query cell to each reference cell type using
Kendall correlation coefficient
}
