% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_cor_mtx}
\alias{get_cor_mtx}
\title{Get correlation matrix}
\usage{
get_cor_mtx(
  sc_count_mtx,
  labels,
  query_mtx,
  global_markers,
  query_nmf_embedding,
  threads = 10,
  verbose = T
)
}
\arguments{
\item{sc_count_mtx}{reference gene counts matrix}

\item{labels}{reference cell type labels}

\item{query_mtx}{query_mtx query gene activity matrix}

\item{global_markers}{global markers got from \code{get_global_markers_sc()} or \code{get_global_markers_bulk()}}

\item{query_nmf_embedding}{query meta-program matrix}

\item{threads}{the number of threads, default is 10}

\item{verbose}{whether to display messages, default is TRUE}
}
\value{
Returns a correlation matrix whose rows are query cells and columns are reference cell types
}
\description{
Get correlation matrix of each query cell to each reference cell type using Kendall correlation coefficient
}
