% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round2Anno.R
\name{get_final_seed_barcodes}
\alias{get_final_seed_barcodes}
\title{get_final_seed_barcodes}
\usage{
get_final_seed_barcodes(
  cell.meta,
  candidate.seed.barcodes,
  query.nmf.embedding,
  k = 10
)
}
\arguments{
\item{cell.meta}{todo}

\item{candidate.seed.barcodes}{todo}

\item{query.nmf.embedding}{todo}
}
\value{
todo
}
\description{
get_final_seed_barcodes
}
\examples{
todo
}
