% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_global_markers_bulk}
\alias{get_global_markers_bulk}
\title{Get global markers for bulk RNA-seq gene counts matrix}
\usage{
get_global_markers_bulk(
  sc_counts_mtx,
  labels,
  max_marker = 200,
  threads = 10,
  return_raw = FALSE
)
}
\arguments{
\item{sc_counts_mtx}{Bulk RNA-seq counts matrix}

\item{labels}{Bulk RNA-seq sample labels}

\item{max_marker}{Maximum number of markers for each cell type}

\item{threads}{The number of threads}

\item{return_raw}{Whether return raw data}
}
\value{
Returns a list of global markers for each cell type
}
\description{
Get global markers for bulk RNA-seq gene counts matrix
}
