% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_global_markers_sc}
\alias{get_global_markers_sc}
\title{Get global markers for scRNA-seq gene counts matrix}
\usage{
get_global_markers_sc(
  sc_counts_mtx,
  labels,
  max_marker = 200,
  threads = 10,
  return_raw = F
)
}
\arguments{
\item{sc_counts_mtx}{scRNA-seq counts matrix}

\item{labels}{scRNA-seq cell labels}

\item{max_marker}{Maximum number of markers for each cell type}

\item{threads}{The number of threads, default is 10}

\item{return_raw}{Whether return raw data, default is FALSE}
}
\value{
Returns a list of global markers for each cell type
}
\description{
Get global markers for scRNA-seq gene counts matrix
}
