% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{get_kendall_pred}
\alias{get_kendall_pred}
\title{Get first round annotation labels}
\usage{
get_kendall_pred(cor_mtx)
}
\arguments{
\item{cor_mtx}{The correlation matrix get from \code{get_cor_mtx()}}
}
\value{
Returns a cell metadata whose rows are query cells and columns are first round labels.
}
\description{
Get first round annotation labels from the correlation matrix
}
