% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.2_MAPLabelCorrector.R
\name{get_merged_labels}
\alias{get_merged_labels}
\title{Get merged labels}
\usage{
get_merged_labels(pred_df)
}
\arguments{
\item{pred_df}{a data frame whose rows are cells and each column is the predictions from a specific reference}
}
\value{
Returns the merged labels
}
\description{
Get merged labels
}
