% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{get_neighbor_celltypes}
\alias{get_neighbor_celltypes}
\title{Get neighbor cell types}
\usage{
get_neighbor_celltypes(
  sc_count_mtx,
  labels,
  global_markers,
  min_cor = 0.6,
  verbose = T
)
}
\arguments{
\item{sc_count_mtx}{scRNA-seq counts matrix}

\item{labels}{scRNA-seq cell labels}

\item{global_markers}{Global markers got from \code{get_global_markers_sc()}}

\item{min_cor}{The minimum cutoff of cell type correlation to define 'neighbor cells'}

\item{verbose}{Whether to display messages and plot}
}
\value{
Returns a list of cell subtypes
}
\description{
Get neighbor cell types from a pseudo bulk matrix
}
