% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.1_WKNN.R
\name{get_neighbors}
\alias{get_neighbors}
\title{get_neighbors}
\usage{
get_neighbors(train, test, k, dist.metric = "cosine")
}
\arguments{
\item{train}{todo}

\item{test}{todo}

\item{k}{todo}

\item{metric}{euclidean, cosine, manhattan, and hamming}
}
\value{
todo
}
\description{
get_neighbors
}
