% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{get_nmf_embedding}
\alias{get_nmf_embedding}
\title{Get meta-program matrix by NMF}
\usage{
get_nmf_embedding(
  peak_counts,
  binarize = TRUE,
  tfidf = TRUE,
  normalize = TRUE,
  nmf_seed = NULL,
  n_factors = 50L,
  verbose = TRUE
)
}
\arguments{
\item{peak_counts}{A query scATAC-seq peak counts whose rows are peaks and columns are cells}

\item{binarize}{Whether to binarize the peak counts matrix}

\item{tfidf}{Whether to do TF-IDF}

\item{normalize}{Whether to do normalization}

\item{nmf_seed}{The seed set for NMF}

\item{n_factors}{The number of factors (meta-programs) to get}

\item{verbose}{Whether to display messages}
}
\value{
Returns a dimension reduced matrix whose rows are cells and columns are factors (meta-programs).
}
\description{
Get meta-program matrix by NMF from scATAC-seq peak counts matrix
}
