% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.1_PreProcessing.R
\name{get_nmf_embedding}
\alias{get_nmf_embedding}
\title{Get dimension reduced cell embedding by NMF}
\usage{
get_nmf_embedding(
  peak_counts,
  binarize = T,
  tfidf = T,
  normalize = T,
  nmf_seed = NULL,
  n_factors = 50L,
  verbose = T
)
}
\arguments{
\item{peak_counts}{a query scATAC-seq peak counts whose rows are peaks and columns are cells}

\item{binarize}{whether to binarize the peak counts matrix, default is TRUE (recommended)}

\item{tfidf}{whether to do TF-IDF, default is TRUE (recommended)}

\item{normalize}{whether to do normalization, default is TRUE (recommended)}

\item{nmf_seed}{the seed set for \code{RcppML::nmf}, default is NULL}

\item{n_factors}{the number of factors (meta-programs) to get, default is 50}

\item{verbose}{whether to display messages, default is TRUE}
}
\value{
Returns a dimension reduced matrix whose rows are cells and columns are
factors (meta-programs).
}
\description{
Get dimension reduced cell embedding by NMF from scATAC-seq peak counts
}
\examples{
query_nmf_embedding <- get_nmf_embedding(PEAK_COUNTS, n_factors = 50)
}
