% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{get_pseudo_bulk_mtx}
\alias{get_pseudo_bulk_mtx}
\title{Get the pseudo bulk matrix}
\usage{
get_pseudo_bulk_mtx(sc.mtx, cell.type, mode = "mean")
}
\arguments{
\item{sc.mtx}{a single cell matrix whose rows are features and columns are cells}

\item{cell.type}{a vector of cell types whose orders are same as the columns of sc.mtx}

\item{mode}{the method to aggragate cells, default is \code{'mean'}
\itemize{
\item{'mean': }{get the mean of each cell type}
\item{'sum': }{get the sum of each cell type}
\item{'median': }{get the median of each cell type}
}}
}
\value{
Returns a pseudo bulk matrix whose rows are features and columns are cell types.
}
\description{
Get the pseudo bulk matrix from a single cell matrix
}
