% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.5_GetSeedCandidates.R
\name{get_seed_candidates}
\alias{get_seed_candidates}
\title{Get seed cell candidates}
\usage{
get_seed_candidates(cell_meta, threads = 10)
}
\arguments{
\item{cell_meta}{a cell metadata}

\item{threads}{the number of threads, default is 10}
}
\value{
Returns a new cell_meta with column \code{GMSS_cutoff} and \code{NMSS_cutoff}
}
\description{
Determine GMSS cutoff and NMSSS cutoff for each cell type
}
