% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{pre_processing}
\alias{pre_processing}
\title{Pre-process input matrices}
\usage{
pre_processing(ref_mtx, query_mtx, verbose = T)
}
\arguments{
\item{ref_mtx}{A reference gene expression matrix whose rows are genes and columns are cells/samples}

\item{query_mtx}{Query gene activity matrix whose rows are genes and columns are cells}

\item{verbose}{Whether to display messages, default is TRUE}
}
\value{
Returns a list containing two matrices after pre-processing.
}
\description{
Pre-process two matrices to remove 0 expressed genes and keep common genes
}
