% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Main.R
\name{RunAtacAnnoR}
\alias{RunAtacAnnoR}
\title{Run AtacAnnoR in a line}
\usage{
RunAtacAnnoR(
  ref_mtx,
  ref_celltype,
  ref_type = "sc",
  query_gene_activity,
  query_peak_counts = NULL,
  query_nmf_embedding = NULL,
  threads = 10,
  verbose = TRUE,
  simple_output = TRUE
)
}
\arguments{
\item{ref_mtx}{a reference gene expression matrix whose rows are genes and columns are cells/samples}

\item{ref_celltype}{a vector of cell types whose orders are same as the columns of sc.mtx}

\item{ref_type}{which type of the reference matrix is
\itemize{
\item{'sc': }{the reference matrix is a single cell RNA-seq matrix}
\item{'bulk': }{the reference matrix is a bulk RNA-seq matrix}
}}

\item{query_gene_activity}{a query gene activity matrix whose rows are genes and columns are cells}

\item{query_peak_counts}{a query scATAC-seq peak counts whose rows are peaks and columns are cells}

\item{query_nmf_embedding}{a query scATAC-seq NMF embedding whose rows are cells and columns are factors,
can be got by \code{get_nmf_embedding()}}

\item{threads}{the number of threads, default is 10}

\item{verbose}{whether to display messages and plots, default is TRUE}

\item{simple_output}{whether to return a simple output, default is TRUE}
}
\value{
If \code{simple_output} is set to TRUE, returns a vector of predicted labels.
If \code{simple_output} is set to FALSE, returns a list object of all outputs.
}
\description{
AtacAnnoR main function
}
