% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{RunAtacAnnoR}
\alias{RunAtacAnnoR}
\title{Run AtacAnnoR in a line}
\usage{
RunAtacAnnoR(
  ref_mtx,
  ref_celltype,
  ref_type = "sc",
  query_gene_activity,
  query_peak_counts = NULL,
  query_cell_embedding = NULL,
  min_cor = 0.6,
  num_global_markers = 200,
  num_neighbor_markers = 200,
  threads = 10,
  verbose = TRUE,
  simple_output = TRUE
)
}
\arguments{
\item{ref_mtx}{A reference gene expression counts matrix whose rows are genes and columns are cells/samples}

\item{ref_celltype}{A vector of cell types whose orders are same as the columns of \code{ref_mtx}}

\item{ref_type}{Which type of the reference matrix is
\itemize{
\item{'sc': }{The reference matrix is a single cell RNA-seq matrix}
\item{'bulk': }{The reference matrix is a bulk RNA-seq matrix}
}}

\item{query_gene_activity}{A query gene activity matrix whose rows are genes and columns are cells}

\item{query_peak_counts}{A query scATAC-seq peak counts matrix whose rows are peaks and columns are cells}

\item{query_cell_embedding}{A query scATAC-seq cell embedding whose rows are cells and columns are factors,
default is NULL, which means getting NMF embedding automatically}

\item{min_cor}{The minimum of correlation to define similar cell types}

\item{num_global_markers}{The number of global markers}

\item{num_neighbor_markers}{The number of neighbor markers}

\item{threads}{The number of threads}

\item{verbose}{Whether to display messages and plots}

\item{simple_output}{Whether to return a simple output}
}
\value{
If \code{simple_output} is set to TRUE, returns a vector of predicted labels.
If \code{simple_output} is set to FALSE, returns a list object of all outputs.
}
\description{
AtacAnnoR main function.
}
