% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{RunAtacAnnoR_ArchR}
\alias{RunAtacAnnoR_ArchR}
\title{Run AtacAnnoR on an ArchRProject in ArchR pipeline}
\usage{
RunAtacAnnoR_ArchR(
  ref_mtx,
  ref_celltype,
  ref_type = "sc",
  query_ArchRproj,
  threads = 10,
  verbose = TRUE
)
}
\arguments{
\item{ref_mtx}{a reference gene expression matrix whose rows are genes and columns are cells/samples}

\item{ref_celltype}{a vector of cell types whose orders are same as the columns of sc.mtx}

\item{ref_type}{which type of the reference matrix is
\itemize{
\item{'sc': }{the reference matrix is a single cell RNS-seq matrix}
\item{'bulk': }{the reference matrix is a bulk RNA-seq matrix}
}}

\item{query_ArchRproj}{the query ArchRProject}

\item{threads}{the number of threads, default is 10}

\item{verbose}{whether to display messages and plots, default is TRUE}
}
\value{
Returns a new query ArchRProject with cell types predicted by AtacAnnoR restored in
\code{query_ArchRproj$final_pred}.
}
\description{
Run AtacAnnoR on an ArchRProject in ArchR pipeline Note that before using this
function, the \code{GeneScoreMatrix} and \code{PeakMatrix} should be stored in
\code{query_ArchRproj}, which can be checked by \code{getAvailableMatrices(query_ArchRproj)}.
}
