% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Main.R
\name{RunAtacAnnoR_Cicero}
\alias{RunAtacAnnoR_Cicero}
\title{Run AtacAnnoR on a cell_data_set(cds) object in Cicero pipeline.}
\usage{
RunAtacAnnoR_Cicero(
  ref_mtx,
  ref_celltype,
  ref_type = "sc",
  query_cds,
  query_gene_activity,
  threads = 10,
  verbose = TRUE
)
}
\arguments{
\item{ref_mtx}{a reference gene expression matrix whose rows are genes and columns are cells/samples}

\item{ref_celltype}{a vector of cell types whose orders are same as the columns of sc.mtx}

\item{ref_type}{which type of the reference matrix is
\itemize{
\item{'sc': }{the reference matrix is a single cell RNS-seq matrix}
\item{'bulk': }{the reference matrix is a bulk RNA-seq matrix}
}}

\item{query_cds}{the query cds object}

\item{query_gene_activity}{a query gene activity matrix whose rows are genes and columns are cells}

\item{threads}{the number of threads, default is 10}

\item{verbose}{whether to display messages and plots, default is TRUE}
}
\value{
Returns a new query cds object with cell types predicted by AtacAnnoR restored in
\code{query_cds$final_pred} and NMF embedding restored in \code{query_cds@int_colData$reducedDims$NMF}.
}
\description{
Run AtacAnnoR on a cell_data_set(cds) object in Cicero pipeline. Note that before using this
function, gene activity should be calculate by Cicero (\code{https://cole-trapnell-lab.github.io/cicero-release/docs_m3/}).
}
