% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Main.R
\name{RunAtacAnnoR_Signac}
\alias{RunAtacAnnoR_Signac}
\title{Run AtacAnnoR on a Seurat object in Signac pipeline}
\usage{
RunAtacAnnoR_Signac(
  ref_SeuratObj,
  ref_assay = "RNA",
  ref_ident = "celltype",
  ref_type = "sc",
  query_SeuratObj,
  query_ga_assay = "ACTIVITY",
  query_peak_assay = "ATAC",
  threads = 10,
  verbose = TRUE
)
}
\arguments{
\item{ref_SeuratObj}{the reference Seurat object}

\item{ref_assay}{the assay of the reference Seurat object to use}

\item{ref_ident}{the identity of the reference Seurat object which contains the reference cell type}

\item{ref_type}{which type of the reference matrix is
\itemize{
\item{'sc': }{the reference matrix is a single cell RNS-seq matrix}
\item{'bulk': }{the reference matrix is a bulk RNA-seq matrix}
}}

\item{query_ga_assay}{the assay containing gene activity of the query Seurat object}

\item{query_peak_assay}{the assay containing peak counts of the query Seurat object}

\item{threads}{the number of threads, default is 10}

\item{verbose}{whether to display messages and plots, default is TRUE}

\item{query}{the query Seurat object}
}
\value{
Returns a new query Seurat object with cell types predicted by AtacAnnoR restored in
\code{query$final_pred} and NMF embedding restored in \code{query[['nmf']]}.
}
\description{
Run AtacAnnoR on a Seurat object in Signac pipeline. Note that before using this
function, the \code{Signac::GeneActivity()} should be
run on the query Seurat object to get the gene activity,
and \code{Seurat::NormalizeData()} should be run on both reference
RNA counts and query gene activity counts.
}
