% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.4_TestMarkers.R
\name{test_markers}
\alias{test_markers}
\title{Test markers of first round annotation}
\usage{
test_markers(
  query_mtx,
  cell_meta,
  global_markers,
  neighbor_markers,
  which_label = "kendall_pred",
  threads = 10,
  verbose = T
)
}
\arguments{
\item{query_mtx}{query gene activity matrix whose rows are genes and columns are cells}

\item{cell_meta}{a cell metadata}

\item{global_markers}{global_markers}

\item{neighbor_markers}{neighbor_markers}

\item{which_label}{which_label}

\item{threads}{the number of threads, default is 10}

\item{verbose}{whether to display messages, default is TRUE}
}
\value{
Returns a new cell metadata with columns 'GMSS','NMSS','NMFC'.
}
\description{
Test markers of first round annotation
}
