% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round1Anno.R
\name{test_markers}
\alias{test_markers}
\title{Test markers}
\usage{
test_markers(
  query_mtx,
  cell_meta,
  global_markers,
  neighbor_markers,
  which_label = "kendall_pred",
  threads = 10,
  verbose = TRUE
)
}
\arguments{
\item{query_mtx}{Query gene activity matrix whose rows are genes and columns are cells}

\item{cell_meta}{A cell metadata}

\item{global_markers}{Global markers}

\item{neighbor_markers}{Neighbor markers}

\item{which_label}{Which label to test}

\item{threads}{The number of threads}

\item{verbose}{Whether to display messages}
}
\value{
Returns a new cell metadata with new columns 'GMSS','NMSS'
}
\description{
Test whether the markers are of high activity in the predicted cells
}
