% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round2Anno.R
\name{weighted_knn}
\alias{weighted_knn}
\title{Weighted KNN}
\usage{
weighted_knn(neighbors, train.labels, type = NULL)
}
\arguments{
\item{neighbors}{Neighbors got from \code{get_neighbors()}}

\item{train.labels}{The labels of cells in the train dataset}

\item{type}{Return type, default is NULL, which means only return labels, \code{prob} means return the predicted probability}
}
\value{
Returns the predicted labels
}
\description{
Apply the weighted KNN algorithm to make the predictions
}
